[balsheet-pnl] Refactor. change monetary conversion function
This is a major internal change for displaying original currency.

Previously each cell would call (get-cell-monetary-fn) or
(get-cell-orig-monetary-fn) to retrieve monetary values.

I've changed to make calls to (get-cell-monetary-fn)
or (convert-curr-fn) to optionally convert to report currency. This
will allow a cell to display multiple monetary values, and each
monetary will optionally be converted to report currency (if price is
available).