Bugfix 796783 - handling invalid date range in "get transaction..."
The GnuCash attempts to get the transactions from the online bank even
though the date range entered is incorrect.

Examples for incorrect user inputs:
1) "to"-date is smaller than "from"-date
2) "to"-date is in the future

In both cases GnuCash tries to get a connection to the online bank.
It takes a while. After that GnuCash shows an error message:
"Error on executing job. Status: error - job contains errors".
The error message is very general and does not say what is wrong.

It is better to check the user inputs *before* the request
will be send to the bank, and to show a error message if the entered date
range is wrong, *without* creating the connection to the online bank.