Add the parent window to gnc_restore_window_size function
When a dialogue is open for the first time, there is no saved geometry
so to locate this dialogue try and use the parent window location by
doing the following...

If the dialogue has a size, align it centrally on the parent window.
If the dialogue has no size, let gtk align it.

This commit breaks the build, the following commits changes all
occurrences of gnc_restore_window_size to include parent window.