Bug 796792 - SaveAs Overwrite dialogue in background and not visible
Firstly the wrong parent was being used for these secondary warning
messages, should be the dialogue and not the main Gnucash application so
these have been changed.

Also some of these secondary warning dialogues still had NULL as the
parent and so the transient parent warning would occur, added parent.