Ignore deprecation warnings from glib
In glib 2.58, g_type_class_add_private is deprecated, which causes gnucash
to fail to build due to -Werror.  To get gnucash building in Ubuntu, we have
turned this particular error back into a warning with
-Wno-error=deprecated-declarations.

This also drops the -Wno-deprecated-register flag, which doesn't appear
to be recognized at all by gcc-8.

https://launchpad.net/ubuntu/+source/gnucash/1:3.2-1build1/+build/15137868