Fix detection of big-endianness on Linux/s390x
Detecting big-endian systems currently relies on some cmake WORDS_BIGENDIAN
define which is failing to be set on big-endian architectures in Debian and
Ubuntu (mips, s390x) for reasons that elude me.  Rather than trying to debug
why this is, just use endian.h, which we can rely on unconditionally on
glibc-based systems.