Bug 796839 - Find transaction won't filter on Date Posted
The results of the find are displayed in a search_ledger but with reset
filter settings. As part of the initial setup of the filter, the
SPLIT_RECONCILE and TRANS_DATE_POSTED parameters are purged from the
query and so if 'Date Posted' was used, all transactions were displayed
so this commit tests for a search_ledger and stops the purge along with
saving the query for subsequent use in the filter dialogue.