Reduce the number of times the register loads - Update
In the previous update the changes to 'double_line mode' and 'register
style' were not catered for so this commit fixes that by using a stored
value 'enable_refresh' in the GncPluginPageRegister. This value is used
to enable/disable the triggering of gnc_ledger_display_refresh command
instead of passing a parameter as these settings are triggered by call
backs. The value is set to FALSE in ...recreate_page before restoring
all settings and subsequently set to TRUE after so normal refreshing
can occur. This reduces the potential number of refreshes on load from
7 to 2.