[net-charts] modify process-datelist to cycle balancelist once
This will deconstruct process-datelist to not call the utility
(gnc:accounts-get-comm-total-*) functions which are still slow,
because they will cycle through the balancelist for each account. In a
large enough report, the balance list may be thousands of entries
long, and we don't want to cycle through them every time.

This commit will loop all so that the balances are cycled once only.