Dummy change to experiment with github PR's
This commit should not make it into upstream. It's meant to evaluate a pure github based workflow and in particular how a "Rebase and commit" PR from upstream to a fork will affect future PR's from that fork back to upstream.

To set up the experiment I started with a fork that lagged behind upstream's maint branch.
Next I used github's PR facility to bring the fork up to date:
* Create a PR from upstream to my fork
* Handle the PR by using "Merge pull request" -> "Rebase and merge" instead of "Merge pull request" -> "Create a merge commit"

I noted this pollutes the fork's history: the rebase alters the committer (not the author) and hence all commit hashes change.

Next I have created this dummy commit and will create a PR for it against upstream to evaluate what will happen.