New hierarchy assistant error
When using the new hierarchy assistant, when you cancel or finish an
error is recorded in the trace file for invalid cast from GtkWindow to
GtkDialog. The cause is when removing the added book options which are
from a window and not a dialog so use 'gnc_options_dialog_destroy'
instead. This also fixes when after saving the file and you go straight
to File->Properties you get a blank book options dialogue.