Do not destroy the 'blank split'
With a register in journal mode and you use the
'move to blank transaction' button which puts the cursor in the
'blank split' and then immediately click on another transaction you
loose the ability to add any more transactions using the blank
transaction as there is no 'blank split'.
Test for the old_split being the 'blank split' before destroying.