Fix adding relative transaction associated files
This mainly affects Windows when you associate a file that is in the
root of the associated path head setting. Relative paths are stored
without a leading / and in there normal form. A function is used to
convert the original entries if required to this format and save
them so that it can be added to a path head with a common function.