Allow the text options widget to use all remaining space
The text option widget used for report options has only about 3 lines
visible currently and will not use any available space in the dialogue.

Change the packing for this widget to expand and fill the remaining
dialogue space. If there are more than one text option widget on a page
then the available space will be shared equally.