Bug 796978 - Deleting a split of same account as register cancels the transaction without warning
When autocomplete is used, one of the splits takes on the reference to
the blank_split and if you delete this split the whole transaction is
removed. This change compares the split that is about to be deleted to
the blank_split and if the same moves the reference to another split.