If you use the autocomplete and cancel other registers are not refreshed
When the autocomplete is used the blank transaction gets the copy of
the originating transaction and if you decide to cancel this any other
open register that had a split in the new transaction is left with
blank rows where the transaction was. Add a test for the pending and
blank transaction being equal and refresh all registers if so.