Fix unused return value error from e4836f3c
Re-introducing the fgets() call in e4836f3c that has been removed in
bf55c30 triggers the "ignoring return value of fgets, declared with
attribute warn_unused_result" error. The even better way is to actually
use the return value and check it for non-NULL because only then
we did a successful fgets call where we can use its result.