[cash-flow][budget] rewrite accounts' sub-accounts constructor
previous version was completely wrong: "for all accounts'
sub-accounts, if each sub-account not in list, add *all* sub-accounts
to accounts" repeat for each sub-account. This led to an unnecessarily
long list of accounts and subaccounts, and the subaccounts were
repeated for *each* subaccount.

new version is simpler, and more efficient. append accounts &
sub-accounts, and delete duplicates.