Initial release of README-hr.win32-bin.txt
Croatian translation of the "README.win32-bin.txt" file

Note:
In the bottom of the english file, there is the following sentence: "This is the GnuCash 2.2.x README file for Windows binary packages."

Maybe it would be better to simply drop this sentence, as it points to an "old" GnuCash version. Or at least drop the information about the version. At present GnuCash is version 3.4.