Fix bug 796987 Import Bills & Invoices: fixing import items only works in special case
Removed dependencies of data field values on previous data rows.
The import takes the header data for an invoice from the first data row, so passing header data to subsequent rows is not useful.
And the previous functionality worked only if the id of the current row was blank, which further limited its usefulness.

Import Bills & Invoices: validate header fields only for the first data row of an invoice

As the function takes the header data from the first data row of an invoice,
it is not necessary to validate the header data in subsequent rows.