Restore the global locale after Guile munges it.
There's no way to change the environment locale on Windows so calling
setlocale(LC_ALL, ""), as guile does in its init routine, reads the
user's Language and Region settings instead of the environment
variables. We save the discovered environment locale and call setlocale
with it again after Guile has initialized.