[test-extras] create-multisplit-transaction
This is *the* general case for any transaction creation.  Rewrite
other transaction creation routines to use it.  All tests still work
unchanged, which confirms this function works well.

This will allow tests to create multisplit transactions, of any
number-of-splits. If the list-of-split's values are not
balanced (i.e. total 0), the engine will create an Imbalance split.

The motivation is to allow creation of complex multisplit
multicommodity transactions eg USD50 + GBP20 (USD25) = EUR66 (USD75)
as well as their prices GBP/USD = 25/20 and EUR/USD = 75/66.

* USD -50
* USD -25 = GBP -20
* USD +75 = EUR +66

This will be useful in creating tests for stock-based reports.