Report: Balance Forecast
This report forecasts the combined balances of the selected accounts
based on the scheduled transactions and plots them on a line graph.

You can set a "reserve" amount, which will draw a red line on the
graph, so you can easily see if your forecast dips below a given value.

There is also a "future minimum" line which shows what the lowest future
balance will be at a given point. I find this useful in conjunction with
the "target" line for planning.