[commodity-utils] simplify gnc:make-exchange-function
Remove assigning exchangelist to exchange-alist.

Removes need to test foreign-amount against 0 - multiplication will
result in 0 anyway. The only functional change is the condition
whereby foreign's commodity is not present in exchangelist; previously
it would return (gnc:make-gnc-monetary domestic 0); now it returns #f.

This function *is* allowed to return #f for invalid or missing prices;
see the (and foreign ...) conditional.