Bug 797204 - Qif import Tradable commodities page
This bug has two elements, the first is that when you have a large
number of tradable commodities you get an assistant page for each one
and if too many you loose site of the assistant buttons. To fix this a
notebook was added to the tradable commodities assistant page with the
individual commodities being added to notebook pages.

The second part was that when editing these an intrusive dialogue was
being raised that displayed the same information as the tooltip which
blocked input. These were removed and the assistant next button enabled
when all entries are valid.