Change assistant-ab-initial.glade so it displays correctly
The text on the first page was squashed as the width could not be
negotiated properly as all the text was set to wrap automaticity and on
subsequent pages there was no wide widgets. Moved the first text row to
a separate widget and not wrap-able so a decent width can be negotiated.