[balsheet-pnl] chart: set report-currency to most used currency
This is a usability upgrade.

This balsheet-pnl report does *not* enforce a report currency; it can
print, process and multilevel-total monetary amounts in their original
currency.

But a linked chart needs one. By default it uses global
report_currency which may not match the book's most used currencies.

This commit will scan the accountlist, retrieve the most common
currency used, and set it as the report-currency for the linked
chart.

The book-main-currency function *may* be upgraded to
report-utilities.scm in the future.

Case example:
* set gnc_default_report_currency to USD
* load book.gnucash with majority accounts in EUR, some in GBP
* run balsheet-pnl.scm - shows EUR and GBP amounts and mixed totals.
* click link to barchart - would not succeed because all amounts
  converted to USD0.00. this commit will ensure barchart's target
  currency is set to EUR thereby showing useful chart.