Add new function named similar to remove class function
The existing function name to add CSS classes for style contexts was
not correct so add a new function gnc_widget_style_context_add_class
named similar to remove function and use it for the existing
'gnc-class-highlight'. Change old function to use this new one with the
old one being removed in master.