Bug 797277 - Assets over time ignores nested liabilities
This is caused by the omission of selected accounts' descendant
accounts during generation of accounts-balances-alist.

The subsequent function account->balance-list was querying the
accounts-balances-alist for subaccounts' (of any account-type)
balances but these were missing.

This commit ensures all selected accounts (whose types depended on the
chart-type), and all their descendant accounts (of any account-type)
were scanned to generate the account-balances-list.