Bug 797301/2 - Sub-account register not seeing changes to sub accounts.
If you have a sub account register open and you create another account
that is a descendant of the top account the transaction will not be
seen as the query used for the sub account register does not include
the new account. Add a check to 'refresh_handler' to check for the same
number of sub accounts, if different recreate query.