Error when cancelling the create a new account in register
When you are asked whether you want to create a new account on the
register, possibly due to a typo and you answer no, the same dialogue
will appear another three times before resetting the account cell to
the original value.

These changes eliminate that by returning you back to the cell with the
invalid entry so that you can amend / cancel or use the dialogue again
to create a new account based on an amended entry.