Remove the registered preference callbacks in  gnc-main-window.c
Add a function to remove all the registered preference callbacks that
were setup in gnc-main-window.c and call it from three places. One is
gnc_mian_window_quit which is used when the application quits and
gnc_main_window_close_page, gnc_main_window_destroy which is used when
separate windows are used for pages like the register and they are
closed.