Fix register tooltip crash
When the register is in double line mode, a tooltip for the transaction
association column can be shown if one is present by hovering the mouse
over the cell. If the register 'cursor' is highlighting any transaction
row this works but if the 'cursor' is on a split and mouse moves to an
association cell with an entry, Gnucash will crash. To fix this use the
SheetBlockStyle from block instead of the GnucashCursor.