Bug 737006 - qif-accounts-map stores account names rather than guids
* with this commit it converts prior to file I/O into/from guid
* it can still read old-school full-account-name
* the in-memory gnc-name is still full-account-name

A dirty hack to fix this bug.

The write route analyses in-memory gnc-name, tries to find an account
with the exact same full-name, and writes its GUID instead.

The read routine tries to find an account with the GUID as string. If
an account is found, it will replace the in-memory GUID with the
account's full-name.