Change use of some vertical box widgets to use a grid
There are several places that use two vertical boxes packed into a
horizontal box to arrange a column of labels and a column of entry
widgets which have been changed to use a grid as it is appropriate.
As part of this removal of some gtkalignment widgets to use left/right
margins instead was used.