Stop duplicate entries for the invoice taxtable cell
When the ledger is loaded, the list for the taxtable combo is created
but as the item_list gui is not initialised yet a temporary list store
is used. This is done twice and hence end up with duplicate entries. To
fix this add the missing gtk_list_store_clear to
gnc_combo_cell_clear_menu to clear the temporary list store when the
item_list is not initialised.