[reports] fix some uses of assoc-ref
assoc-ref (or better, assq-ref because we're comparing symbols) is
used to lookup from the 'car's of a list of pairs, and return the
found pair's 'cdr'. the previous use of 2-element lists demonstrate a
lack of understanding how to encode the list of pairs. rewrite using
proper pairs rather than 2-element lists, which means the assq-ref
does not need a subsequent 'car' to retrieve the desired symbol.